Fix a bug causing messages to disappear. DOH. Please upgrade.
Side node: I must have been blind when I wrote that code. Especially it is interesting that I added the assert() statement afterwards to ensure that the db4o query does what I want it to do - and did not notice that the code was broken! FURTHER, it seems that my Eclipse debugger will not break even though that assert() fails and I configured it to break on AssertionFailed. Maybe the ant build script builds the jar in a way which leaves the assert() statements out?

git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/Freetalk@25822 67a373e5-eb02-0410-a15c-ee090a768436