- Adapt the MessageManager to the new SubscribedBoard code: * Upon identity deletion, delete all their messages from subscribed boards. When an OwnIdentity is deleted, delete it's subscribed boards. *  Add code for subscribing/unsubscribing - Remove obsolete, commented out code.
There is still some work missing: The download decision for messages is not based on whether a subscription exists yet. Further, if all messages of a board were downloaded already and a new own identity subscribes to it, they will not be added to the board. I'll implement code for that.