WoTMessageFetcher.onGeneratedMetadata(): Fix to adhere to specification and free the metadata
The callback isn't actually used in this class yet - it's only used for
inserts which this class doesn't do - but it's a good idea to adhere to
the interface specification nevertheless to be prepared for future
changes.
As you can see in the previous commit which touches all other
implementations of this callback this is the way all the other ones
handle it as well.