Fix compiler errors of giving wrong arguments to HighLevelSimpleClient.insert()
Errors would be similar to "The method insert(InsertBlock, String,
boolean, InsertContext, ClientPutCallback, short) in the type
HighLevelSimpleClient is not applicable for the arguments (InsertBlock,
boolean, null, boolean, InsertContext, IdentityInserter, short)".

The getCHKOnly boolean parameter was moved to InsertContext by fred
commit:
https://github.com/freenet/fred/commit/6596c3e238340ce19a34f9e32611862c8f2a4390

It defaults to false so there's no need to manually set it on the
InsertContext.