MessageReference: Change type of mURI from FreenetURI to String
FreenetURI is a fred class, and db4o was removed from fred.
Thus its functions to store it inside db4o databases were removed and we
need to store it as a String.

Also renaming the field to mFreenetURI as the URI type isn't obvious
anymore by changing the class.

Removing the null-checks in storeWithoutCommit() / deleteWithoutCommit()
as the field is final and thus cannot be nulled.
It was likely being null because mFreenetURI was a non-db4o-primitive
class previously and thus could be deleted from the database by
bogus calls removeFrom(), which would even null the final field.
However to me it looks like that bug cannot exist anymore, it was
probably fixed some time ago but I forgot to remove the null-checks.

Not writing code for updating old databases since commit
e2998f20e38d3faf38ebbb35d6d9f469143dc333 prevents loading of them.