Travis CI: Fix to not continue compilation if a command fails
While even without this the build would fail, it would still continue
compiling and only fail at the end.
This is a waste of time, and I think it might cause problems once we add
code to upload the Freetalk JAR to Freenet - the last thing we should
want is a JAR being uploaded even though it is broken.