Travis CI: Remove hack which was copy-pasted from fred's Travis
It's not our job here to optimize the compilation of fred, if fred
developers think this is mandatory then they should be add it to their
gradle.properties file in Git, not manually apply it by Travis shell
scripting.
FWIW Freetalk doesn't run the fred unit tests anyway, so it likely
doesn't matter that we don't do this.