Add db4o-7.4 as git submodule
In prepration of its upcoming removal from fred's freenet-ext.jar

This was done using:
	git submodule add https://github.com/xor-freenet/db4o-7.4.git
	( cd db4o-7.4 ; git verify-tag "$(git describe)" )
	git commit --gpg-sign

Notice that this hardcodes the used revision of that repository to
commit 831aecf5f88a8f6716707c61249d55d627246fac (= tag v34):
If you do "git show" on this commit here which added the submodule you
can see that the commit hash of db4o is hardcoded.
Changing it would require updating the submodule *and* committing that
update of the submodule revision to the Freetalk repository.

Thus anyone who clones Freetalk and downloads the submodule will get
that commit, not a later one. This means that I cannot maliciously alter
db4o in the future :) - which in turn allows Freenet to not have to host
a copy of the db4o repository.

Source: https://git-scm.com/book/en/v2/Git-Tools-Submodules

To obtain db4o after cloning Freetalk use:
	git submodule update --init
NOTICE: The following commits will amend the Ant builder to do this
automatically, and also to actually include it during compilation.