Ant: Use and ship db4o submodule
- Use db4o.jar in classpath during compilation and unit tests
- Package its classes into Freetalk.jar

I've tested this with a freenet-ext.jar which I've modified to not include
db4o and the unit tests do succeed.

The non-unified style of this is intentional:
This was produced by merging changes to WoT's build.xml, ignoring style
simplifies getting all changes merged as it ensures the diff will be
as empty as possible once all improvements from WoT have been merged.