README: Copy section "Compiling with Eclipse" from WoT README to adapt it to Gradle
Not adapted to Freetalk yet, the following commits will deal with that.

Copied from WoT by producing a patch via:
> git diff c75ddaf561814ac2e47ec657f3a5d75ae1847e50..03648c47222df60e2631d787b4f3a74ed949e933
These are the changes of the branch 'wire-gradle-in-to-eclipse-2', one
purpose of which was to adapt the README's Eclipse section to Gradle.
Non-README changes have not been applied here yet, they will ship
separately.