Gradle: Add directory to generate file "Version.properties" into
This file will be consumed by the new Version.getGitRevision()
implementation which the previous branch added.

The proper way to generate resources, which this is inspired by, is
documented in the section "Working with generated resources" of the
following link - notice that the cited section title lacks marking by
e.g. bold as such, it appears as regular text of the site:
https://docs.gradle.org/current/dsl/org.gradle.api.tasks.SourceSetOutput.html