GitHub Actions: Remove Java 9 & 10, fred compilation fails there and it's probably not possible to fix
It fails to download fred's dependencies because Gradle cannot find a
suitable SSL certificate trust path, see the second to last commit's
message.

I suspect this is because Java 9 and 10 are EOL so the certificates they
ship aren't updated anymore in the Adopt OpenJDK distribution which GHA
uses.
It probably did work on Travis CI because it might have used a different
OpenJDK distribution, or because the certificates expired just some days
ago.

Either way this is not worth the effort trying to fix it any more
because supporting EOL Java versions is not our job.