Gradle: Add Maven Central for dependency download if environment variable is set
This will be used for GitHub Actions:
On Windows and MacOS we cannot obtain JUnit etc. from /usr/share so we
need to download the dependencies there.

Not using jcenter() like fred does because jcenter was abandoned and
hence isn't updated anymore:
https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/

mavenCentral() seems to be the default suggestion according to the
Gradle documentation:
https://docs.gradle.org/current/userguide/declaring_repositories.html