Gradle: Logging: Don't imply we're going to support gradle-witness
I do not plan to spend time on it:
Using the binaries from /usr/share is the default for Freetalk.
And that is a good choice because those come from the Linux
distribution, so the distribution can deal with figuring out how to
avoid supply chain attacks - which totally makes sense, because
"distribution" means precisely what this is about, *distribution* of
binaries in a safe fashion.