GitHub Actions: cron: Try to fix "Error: Can't find 'action.yml' [...]"
For why this would happen see the hereby removed FIXME.

The hereby removed checkout of the Freetalk git repository was an earlier
workaround attempt for the issue.
As it did not fix it, it is removed.

It is likely that this commit will yield failure as well, but for a
different reason:
AFAIK files are not preserved across different "job:" sections, so the
FIXME which is kept hereby must be resolved for things to work.