GitHub Actions: README: Add tooltip text for GHA status badge image
Copied from the alt text at the beginning of the line.
(The alt text is what is displayed if the image doesn't load or a blind
person uses a screenreader. The tooltip on the other hand is displayed
on mouseover.)

Based on WoT commit 6aeeea11386c47e1168586865193aa5a7c3fd3ee