Revert "GitHub Actions: README: Add badge for CI test results on Freenet's repository"
This reverts commit f27898cddacfd41e14f9957d209975585e03e09b.

I've reconsidered the whole choice of badges:

With Travis CI, the badges used to show the result of the last *cronjob*
run. That's nice because it shows whether the tests work against
whatever is the most recent OS environment on Travis.

With GHA, by default the result is instead of the run triggered by the
last push.
This might be months or even years outdated!
The changed OS environment might have broken the build already.

Hence it is not a good idea to show this as the badge.

Instead we should show the badge of the separate cronjob workflow which
I've already added to compensate for the lack of cronjobs for "default"
workflows on GHA.
The next commit will do that.

Since the cronjob will only run at my (xor-frenet) Freetalk repo, not at
the one of Freenet, the badge for the Freenet repo is hereby removed.