GitHub Actions: CI tests: Fix fred cache key to include Java version
This ensures we use fred JARs which were compiled using the specific
Java version which we're testing Freetalk against when using the cache.

In theory it shouldn't matter with which Java version fred was compiled as
long as it is above the minimum one, but in favor of defensive programming
I'd rather treat it like a black box and not make such assumptions.