Gradle: Try to fix "Adding a Configuration as a dependency is no longer allowed as of Gradle 8.0."
The hereby removed line caused Gradle to fail with the cited error.

I'm not sure if it is even still necessary to manually tell Gradle that
the testImplementation depends on the junit dependencies, which is what
the line was supposed to do.

So let's just try if it figures that out on its own without the line.