GitHub Actions: Windows: Don't use fred's old Gradle but use the one GHA ships in order to fix "Could not set unknown property 'archiveBaseName' for task ':jar' of type org.gradle.api.tasks.bundling.Jar."
The above error is printed by Gradle because the previous branch changed
the Gradle build script to use the proptery "archiveBaseName" instead of
"baseName".
That was necessary because Gradle 8.0, which we use on non-Windows test
runs, renames the property to "archiveBaseName" and would fail with the
old name.

In other words: fred's Gradle is too old for the current state of
Freetalk's Gradle build script.