compare-gradle-jars-with-ant-jars: Exit with error on MacOS, the script does not work there and fixing it would be too much work
The point of the script is to check if the Gradle builder produces the
same output as the Ant builder.
As long as we can validate that on one OS - Linux - that's probably
enough because the builders don't contain any conditions which check
which OS is being used to do different things based on that.