Revert "GitHub Actions: Amend the previous commit" Revert "GitHub Actions: Try to fix failure on recent Java versions due to fred"
This reverts commit d04c43ef5668b889691c63abce20fcc7cc5f573f.
This reverts commit 356b6c59bfb71432d6d2c91adfa3c9da7fd7d33c.

fred's Gradle has been updated, it now compiles on Java 22.
So we don't need those workarounds anymore.

As seen in the removed comments, I had considered keeping the logic to
use the cached fred JARs which were built on the minimum Java version
which fred needs for building (currently Java 8).
I'ved decided against that because:
- it is not deterministic as explained in the removed comments.
- fred developers do not test whether fred compiles on recent Java
  versions, so let's just test it here as a side effect.